requirements.txt
# Webフレームワーク
flask==3.1.1
gunicorn==21.2.0

# 画像処理
pillow==9.0.1

# 数値計算
numpy==2.2.6
